ai/pyproject.toml
[tool.poetry]
name = "tilly"
version = "0.1.0"
description = "Anomaly detection applied to room usage"
authors = ["nthomsencph <nicolai.thomsen@nttdata.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
pandas = "^2.0.3"
numpy = "^1.25.2"
scikit-learn = "^1.3.0"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
loguru = "^0.7.2"
python-decouple = "^3.8"
databases = {extras = ["sqlite"], version = "^0.8.0"}
snowflake-sqlalchemy = "^1.5.0"
fastapi-users = {version = "<12.1.2", extras = ["sqlalchemy"]}
gunicorn = "^21.2.0"
plotly = "^5.17.0"
jinja2 = "^3.1.2"
snowflake-snowpark-python = {extras = ["pandas"], version = "^1.8.0"}
tqdm = "^4.66.1"
rich = "^13.6.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.4.3"
mkdocstrings = {extras = ["python"], version = "^0.23.0"}

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.0"
jupyter = "^1.0.0"
matplotlib = "^3.7.2"
pre-commit = "^3.3.3"
httpx = "^0.25.0"
pytest = "^7.4.2"
fastapi-profiler = "^1.2.0"

[tool.vulture]
make_whitelist = true
min_confidence = 80
paths = ["tilly"]
sort_by_size = true
verbose = true

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
